requirements.txt
pandas
openpyxl
python-pptx
python-docx
plyer
pyinstaller
